Updated fixture Card to display home and away team emblems as well as match date and time.

+++ app/src/main/java/com/example/fcmilan/models/fixtures/Fixture.kt
@@ -9,7 +9,7 @@ data class Fixture (
   @SerializedName("referee"   ) var referee   : String?  = null,
   @SerializedName("timezone"  ) var timezone  : String?  = null,
   @SerializedName("date"      ) var date      : String?  = null,
-  @SerializedName("timestamp" ) var timestamp : Int?     = null,
+  @SerializedName("timestamp" ) var timestamp : Long?     = null,
   @SerializedName("periods"   ) var periods   : Periods? = Periods(),
   @SerializedName("venue"     ) var venue     : Venue?   = Venue(),
   @SerializedName("status"    ) var status    : Status?  = Status()
